Continued work on Entity
Added a new behavior 'PLAYER' to EnumBehavior. Added methods to AABB that allow for it both moved and modified. Refactored EnumSide so that the Enum is called 'EnumSide' and changed it to a unsigned char from an int. Added a few opperator overloads to Position. Changed TilePosition so that its Position can be changed. Added several methods to Entity that allow for checking states and modifcation of the class.