Worked on MapManager's rendering system
Made rendering universal, previously rendering had been relative to Tile-Space where as now it is relative to Global-Space. This means that MapManager will now calculate the number of tiles on screen itself instead of the calculation being done inside of GameState. It will also transform the camera coordinate system from Global to Tile-Space. The drawCustom method inside of GameState will now be given both the Camera position and Window size. These arguments should be used instead of calling the Game object when rendering as they been modifided to stay inside of the screen.