Worked on EntityManager Entity vs Tile Collision
Removed all referances to Game_Dynamic in BlankGameState. Corrected an issue with IEntity's AABB not getting spawned in the correct place. Corrected an issue with the way movement is hanndled in IEntity. Made correction to RectEdge in Collisions, before the change it would not return the correct side that had been collided with. Added class GameCamera that will hanndle the camera.